feat(session): persist tiled order across restarts; fix zoom_to_fit id translation
Session format v3 records each monitor's monitor_clients sequence as
ordered class+instance identities. restore_session replays the saved
order before the unified arrange via a pure planner: occurrence-aware
matching, vanished identities skipped, new windows appended in their
current relative order, and the floating/tiled grouping invariant wins
over the saved sequence. v1/v2 files migrate in memory; a v3 fixture
freezes the format. The minimized dock's order already rides the
per-window restore property and stays out of the session file.

compositor_zoom_to_fit now crosses the backend boundary as an
Option<WindowId> translated at each delegation site, instead of a raw
u32 that the X11 compositor happened to key by XID; the Wayland side
stores its native u64. No callers exist yet, but the next wiring can no
longer repeat the id-mistranslation bug class the tab bar once had.

Docs catch up with the day's interaction work: new window-tabs, expose
and window-switcher pages (the latter flags the Alt+Tab default change
and shows how to rebind loopview), and updates to notifications,
ui-theme, launcher, minimized-dock, architecture and README.

+++ src/backend/wayland_udev/compositor/config.rs
@@ -1773,7 +1773,7 @@ impl WaylandCompositor {
         self.annotation_line_width = width.max(1.0);
     }
 
-    pub(crate) fn zoom_to_fit(&mut self, window: Option<u32>) {
+    pub(crate) fn zoom_to_fit(&mut self, window: Option<u64>) {
         self.zoom_to_fit_window = window;
         self.needs_render = true;
     }
